Czinger 21C Price: .7 Million
You might not know the name Czinger yet, but this California-based boutique automaker is responsible for the .7 million 21C supercar. With 1,250 horsepower on tap thanks to a 2.9-liter twin-turbocharged V8, the 21C can rocket to 60 miles per hour in 1.9 seconds and a top speed of around 281 miles per hour.

push pull to firebase